Fix: common page 오류 수정함
